all_set/pyproject.toml
[tool.poetry]
name = "all-set"
version = "0.1.0"
description = ""
authors = ["ajuijas <ijas.ahmd.ap@gmail.com>"]
readme = "README.md"
packages = [{include = "all_set"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.3"
djangorestframework = "^3.15.2"
gtts = "^2.5.4"
speechrecognition = "^3.11.0"
pypdf2 = "^3.0.1"
openai = "^1.55.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
